news_parser_api/requirements.txt
beautifulsoup4==4.9.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.1
Flask-HTTPAuth==3.3.0
geographiclib==1.50
geopy==1.21.0
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.1
joblib==1.0.0
lxml==4.6.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.5
Pillow==8.1.0
python-dateutil==2.8.1
python-dotenv==0.12.0
PyYAML==5.4.1
regex==2020.11.13
requests==2.23.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.56.0
urllib3==1.25.8
Werkzeug==1.0.0
